Safari-609.1.11.1/LayoutTests/platform/wpe/css3/flexbox/minimum-size-image-expected.txt


PASS .flexbox 1 
PASS .flexbox 2 
PASS .flexbox 3 
PASS .flexbox 4 
FAIL .flexbox 5 assert_equals: 
<div class="flexbox column" style="height: 10px;" data-expected-height="10">
    <!-- should use content height, clamped by converted max-width, as minimum height. -->
    <img src="../../images/resources/green-100.png" style="max-width: 5px;" data-expected-width="5" data-expected-height="5">
</div>
height expected 5 but got 4
PASS .flexbox 6 
PASS .flexbox 7 
FAIL .flexbox 8 assert_equals: 
<div class="flexbox" style="width: 10px;" data-expected-width="10">
    <!-- should use content width, clamped by converted min-height, as minimum width. -->
    <img src="../../images/resources/green-100.png" style="max-height: 5px;" data-expected-width="5" data-expected-height="5">
</div>
width expected 5 but got 4
PASS .flexbox 9 
FAIL .flexbox 10 assert_equals: 
<div class="flexbox" style="width: 10px;" data-expected-width="10">
    <!-- should use content width, clamped by converted max-height, as minimum width. -->
    <img src="../../images/resources/green-100.png" style="max-height: 5px;" data-expected-width="5" data-expected-height="5">
</div>
width expected 5 but got 4
